--- a/pandas/core/indexes/api.py
+++ b/pandas/core/indexes/api.py
@@ -239,8 +239,12 @@ def union_indexes(indexes, sort: bool | None = True) -> Index:
         Index
         """
         if all(isinstance(ind, Index) for ind in inds):
-            result = inds[0].append(inds[1:]).unique()
-            result = result.astype(dtype, copy=False)
+            inds = [ind.astype(dtype, copy=False) for ind in inds]
+            result = inds[0].unique()
+            other = inds[1].append(inds[2:])
+            diff = other[result.get_indexer_for(other) == -1]
+            if len(diff):
+                result = result.append(diff.unique())
             if sort:
                 result = result.sort_values()
             return result